# src/swagger/swaggerDocument.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Eco-Shoop OpenApi specification",
    "version": "0.1.1"
  },
  "servers": [
    {
      "url": "https://eco-food-backend-production.up.railway.app/api/",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3030/api/",
      "description": "Dev server"
    }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Method to get a list of all products",
        "tags": [
          "Products"
        ],
        "operationId": "getAllProducts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of products per page (default: 20)",
            "schema": {
              "type": "number"
            },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_url}": {
      "get": {
        "summary": "Method to get one product by url",
        "tags": [
          "Products"
        ],
        "operationId": "getOneProduct",
        "parameters": [
          {
            "name": "product_url",
            "in": "path",
            "required": true,
            "description": "Url of Product",
            "schema": {
              "type": "string"
            },
            "example": "dried-apricots-elegant"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneProduct"
                }
              }
            }
          },
          "404": {
            "description": "Category or product not found"
          }
        }
      }
    },
    "/products/by-category/{category_url}": {
      "get": {
        "summary": "Method for getting a products list in a category",
        "tags": [
          "Products"
        ],
        "operationId": "getProductsByCategory",
        "parameters": [
          {
            "name": "category_url",
            "in": "path",
            "required": true,
            "description": "Url of Category",
            "schema": {
              "type": "string"
            },
            "example": "dried-fruits"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of products per page (default: 20)",
            "schema": {
              "type": "number"
            },
            "example": 20
          },
          {
            "name": "country",
            "in": "query",
            "description": "Country ID (used to filter products by country) for multiple selection use \"_\" (6463cc86a1290a70677fa0f5_642dda4fe42ee57de3e9e0ca)",
            "schema": {
              "type": "string"
            },
            "example": "645f961a434c323f287affa5"
          },
          {
            "name": "availability",
            "in": "query",
            "description": "Availability ID (used to filter products by availability) for multiple selection use \"_\"",
            "schema": {
              "type": "string"
            },
            "example": "6463cc86a1290a70677fa0f5"
          },
          {
            "name": "brand",
            "in": "query",
            "description": "Brand ID (used to filter products by brand) for multiple selection use \"_\"",
            "schema": {
              "type": "string"
            },
            "example": "645bb6b7a5fe761dbb3f6f68"
          },
          {
            "name": "form",
            "in": "query",
            "description": "Form ID (used to filter products by packaging) for multiple selection use \"_\"",
            "schema": {
              "type": "string"
            },
            "example": "6463a40ddbcd4375578ae45b"
          },
          {
            "name": "price",
            "in": "query",
            "description": "Minimum and maximum price of products (used to filter products by price)",
            "schema": {
              "type": "string"
            },
            "example": "4_11"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "product sorting options (relevancy, price-up, price-down, rating)",
            "schema": {
              "type": "string"
            },
            "example": "relevancy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductSchema": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "642dda4fe42ee57de3e9e0ca"
          },
          "name": {
            "type": "string",
            "example": "Dried Apricots - Elegant"
          },
          "url": {
            "type": "string",
            "example": "dried-apricots-elegant"
          },
          "price": {
            "type": "number",
            "example": 11
          },
          "oldPrice": {
            "type": "number",
            "example": 12.35
          },
          "size": {
            "type": "number",
            "example": 500
          },
          "availability": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "6463cc86a1290a70677fa0f5"
              },
              "name": {
                "type": "string",
                "example": "in stock"
              }
            }
          },
          "description": {
            "type": "string",
            "example": "Product description"
          },
          "composition": {
            "type": "string",
            "example": "Dried apricots"
          },
          "vitamins": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "A, E, C, B3, B5, B6, β"
            }
          },
          "minerals": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Si, Co, Cr, Mg, Cu"
            }
          },
          "nutritionalValue": {
            "type": "number",
            "example": 241
          },
          "image": {
            "type": "string",
            "example": "https://cdn.shopify.com/collections/image.png"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "dried-fruits"
            }
          },
          "brand": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "645bb6b7a5fe761dbb3f6f68"
              },
              "name": {
                "type": "string",
                "example": "Ayoub`s"
              }
            }
          },
          "country": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "645f961a434c323f287affa5"
              },
              "name": {
                "type": "string",
                "example": "Canada"
              },
              "code": {
                "type": "string",
                "example": "CA"
              }
            }
          },
          "form": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "6463a3f0dbcd4375578ae459"
              },
              "name": {
                "type": "string",
                "example": "prepackaged"
              }
            }
          },
          "delivery": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "646795c4acc8351a33526882"
              },
              "description": {
                "type": "string",
                "example": "delivery 1-2 days"
              }
            }
          },
          "rating": {
            "type": "number",
            "example": 4.5
          }
        }
      },
      "OneProduct": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductSchema"
          },
          "reviews": {
            "$ref": "#/components/schemas/Reviews"
          }
        }
      },
      "Products": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductSchema"
            }
          },
          "filter": {
            "$ref": "#/components/schemas/Filter"
          }
        }
      },
      "Filter": {
        "type": "object",
        "properties": {
          "price": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number",
                "example": 4.95
              },
              "max": {
                "type": "number",
                "example": 14
              }
            }
          },
          "brands": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "645bb587a5fe761dbb3f6f64"
                },
                "name": {
                  "type": "string",
                  "example": "The Nut Market, Ayoub`s"
                }
              }
            }
          },
          "countrys": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "645f964f434c323f287affa9"
                },
                "name": {
                  "type": "string",
                  "example": "Canada"
                },
                "code": {
                  "type": "string",
                  "example": "CA"
                }
              }
            }
          },
          "form": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "6463a40ddbcd4375578ae45b"
                },
                "name": {
                  "type": "string",
                  "example": "on weight"
                }
              }
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "6463ccc0a1290a70677fa0f9"
                },
                "name": {
                  "type": "string",
                  "example": "in stock"
                }
              }
            }
          },
          "delivery": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "646795ebacc8351a33526884"
                },
                "description": {
                  "type": "string",
                  "example": "delivery 1-2 days"
                }
              }
            }
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "646d26d025baf5934fc442ab"
          },
          "text": {
            "type": "string",
            "example": "Some product review"
          },
          "score": {
            "type": "number",
            "example": 4
          },
          "owner": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "64411b97a5f7c71fbe6c24dc"
              },
              "name": {
                "type": "string",
                "example": "User"
              }
            }
          }
        }
      },
      "Reviews": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Review"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}